requirements.txt
requests
Pillow
python-dotenv
flask
Flask-PyMongo
ultralytics